By default, compile from code rather than use pre-compiled jar
The current precompiled Hadoop-BAM jars are built for Hadoop 1.x and
won't work with Hadoop 2.0, though the source code should be
compatible.  This commit changes the default action to fetch the source
code and build it.  The desired revision is set in the attributes
(currently the default is 5.1, but one could set the attribute to any
valid revision).  The previous default recipe has been renamed
"precompiled".